A mechanism to register extra maven caches
By default, loading of Maven plugins will look in a cache directory
maven-cache.gate alongside each xgapp currently on the persistence
stack.  This commit adds a global list of extra caches managed by
gate.util.maven.Utils, which will be searched after the persistence
caches but before the remote repositories.  This provides a way for
an embedded app to ship a cache of the plugins it uses so the app
can operate without an internet connection.