Saving and loading of known/load always plugins config
Re-instated use of the Known_plugin_path and Load_plugin_path attributes in user config.  Directory-based plugins are stored as URLs as before, Maven plugins are stored as [group:artifact:version], and plugins that are in the "default" list are not stored in Known_plugin_path (in the same way as those old-style plugins under GATE_HOME/plugins were not stored in previous versions).