Configuration to build the .jar and .exe installer
The installer-building plugins are in a non-default profile because (a) we don't want to have to build the installers when working in development and (b) the installer bundles tao.pdf, and requires a -D on the command line telling it where to look for that file.

Note that this module now depends on a SNAPSHOT version of izpack (because installers created with the latest release version hang on Java 9 or later), but this shouldn't matter as we never release the artifacts from this module into Central.  Once the next release of izpack comes out with the fix then we can depend on that instead.