Special case for the final segment
When we're asking for the original position of a point at the end boundary of a segment and afterChar==false then we take the start of the next segment instead of the end of this one.  But if the segment in question is the very last one in the RepositioningInfo then there is no "next segment", which was previously causing getOriginalPos to fail (returning -1).  Now if there's no next segment it falls back to using the end of the last segment as it would if afterChar==true.

Closes #79